refactor: Move layer loading progress functionality to separate service
